+++ modules/Accounting/Incomes.php
@@ -102,7 +102,7 @@ if ( ! $_REQUEST['modfunc'] )
 
 	if ( ! empty( $RET ) && ! $_REQUEST['print_statements'] && AllowEdit() && ! isset( $_REQUEST['_ROSARIO_PDF'] ) )
 	{
-		$columns = array( 'REMOVE' => '' );
+		$columns = array( 'REMOVE' => '<span class="a11y-hidden">' . _( 'Delete' ) . '</span>' );
 	}
 	else
 	{