+++ modules/Accounting/StaffPayments.php
@@ -119,7 +119,7 @@ if ( UserStaffID() && ! $_REQUEST['modfunc'] )
 
 	if ( ! empty( $RET ) && ! $_REQUEST['print_statements'] && AllowEdit() )
 	{
-		$columns = array( 'REMOVE' => '' );
+		$columns = array( 'REMOVE' => '<span class="a11y-hidden">' . _( 'Delete' ) . '</span>' );
 	}
 	else
 	{