+++ modules/School_Setup/includes/Modules.inc.php
@@ -198,7 +198,11 @@ if ( ! $_REQUEST['modfunc'] )
 		$modules_RET[] = $THIS_RET;
 	}
 
-	$columns = array( 'DELETE' => '', 'TITLE' => _( 'Title' ), 'ACTIVATED' => _( 'Activated' ) );
+	$columns = array(
+		'DELETE' => '<span class="a11y-hidden">' . _( 'Delete' ) . '</span>',
+		'TITLE' => _( 'Title' ),
+		'ACTIVATED' => _( 'Activated' ),
+	);
 
 	unset( $modules_RET[0] );